requirements.txt
Flask==1.1.2
twilio==6.40.0
python-dotenv==v0.13.0
requests==2.23.0
boto3==1.13.18
